dev/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "ceramicspeed"
version = "0.1.0"
description = "Ball bearing ultrasound & acoustic emission signal analysis pipeline"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy",
    "pandas",
    "scipy",
    "antropy",
    "h5py",
    "pyarrow",
    "joblib",
    "tqdm",
    "pyyaml",
    "click",
    "scikit-learn",
    "matplotlib",
    "plotly",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "ruff",
    "mypy",
]

[project.scripts]
ceramicspeed = "ceramicspeed.cli.main:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "N", "ANN"]
ignore = ["ANN101", "ANN102"]

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true
